+++ app/javascript/pages/Signin.js
@@ -1,0 +1,7 @@
+import React from 'react'
+
+function Signin () {
+  return <div>Signin</div>
+}
+
+export default Signin
